Fixed issue and added more comments for file upload

+++ forum/Controllers/DashboardController.cs
@@ -56,14 +56,14 @@ public class DashBoardController : Controller
 
         return View(userActivity);
     }
-    
+
     // Method for fetching the admin dashboard view
     [Authorize(Roles = "Admin")]
     public async Task<IActionResult> AdminDashboard()
     {
         // Initialize categories and tags
-        IEnumerable<Category>? categories = null;
-        IEnumerable<Tag>? tags = null;
+        IEnumerable<Category>? categories;
+        IEnumerable<Tag>? tags;
 
         try
         {
@@ -74,18 +74,19 @@ public class DashBoardController : Controller
         catch (Exception e)
         {
             // Exception and error logging if the server can't fetch the data
-            _logger.LogError("[Dashboard controller] An exception occurred while fetching categories or tags: {e}", e.Message);
+            _logger.LogError("[Dashboard controller] An exception occurred while fetching categories or tags: {e}",
+                e.Message);
             return StatusCode(500, "Internal server error. Please try again later.");
         }
-        
+
         // Exception and error logging if there are no tags or categories to show the user
         if (categories == null || tags == null)
         {
             _logger.LogError(
                 "[Dashboard controller] _categoryRepository.GetAll() and/or _tagsRepository.GetAll() returned null");
-            return NotFound("Categories or tags not found, cannot create post"); 
+            return NotFound("Categories or tags not found, cannot create post");
         }
-        
+
         // New view model for creating a post
         var adminDashboardViewModel = new DashboardViewModel
         {
@@ -96,10 +97,11 @@ public class DashBoardController : Controller
         return View(adminDashboardViewModel);
     }
 
-     [HttpPost]
+    [HttpPost]
     [Authorize(Roles = "Admin")]
     public async Task<IActionResult> UpdateCategory(Category category)
     {
+        // Get the category from the database
         var dbCategory = await _categoryRepository.GetTById(category.CategoryId);
 
         if (dbCategory == null)
@@ -113,12 +115,11 @@ public class DashBoardController : Controller
         string newPicturePath = "";
 
 
-        Console.WriteLine("Path to delete" + pictureDeletePath);
-
+        var file = Request.Form.Files.FirstOrDefault();
         // If the user has selected a file
-        if (Request.Form.Files.Count > 0) // If the user has selected a file
+        if (file != null) // If the user has selected a file
         {
-            newPicturePath = await FileUpload(category);
+            newPicturePath = await FileUpload(file);
 
             if (newPicturePath.IsNullOrEmpty())
             {
@@ -137,9 +138,11 @@ public class DashBoardController : Controller
         dbCategory.Name = category.Name;
         dbCategory.Color = category.Color;
 
+
         // Update the picture path if the user has used the text field
         if (!category.PicturePath.IsNullOrEmpty())
         {
+            Console.WriteLine("Using text field");
             dbCategory.PicturePath = category.PicturePath;
         }
 
@@ -154,9 +157,11 @@ public class DashBoardController : Controller
         // Update the category
         await _categoryRepository.Update(dbCategory);
 
+
         // Delete the old picture if it exists
-        if (!pictureDeletePath.IsNullOrEmpty())
+        if (!newPicturePath.IsNullOrEmpty())
         {
+            Console.WriteLine("DELETING PICTURE!");
             if (!DeleteFile(pictureDeletePath))
             {
                 _logger.LogError(
@@ -173,10 +178,12 @@ public class DashBoardController : Controller
     [Authorize(Roles = "Admin")]
     public async Task<IActionResult> NewCategory(Category category)
     {
+        var file = Request.Form.Files.FirstOrDefault();
+
         // If the user has selected a file
-        if (Request.Form.Files.Count > 0) // If the user has selected a file
+        if (file != null)
         {
-            string filePath = await FileUpload(category);
+            string filePath = await FileUpload(file);
 
             if (filePath.IsNullOrEmpty())
             {
@@ -193,9 +200,9 @@ public class DashBoardController : Controller
 
         if (ModelState.IsValid)
         {
-            var result = await _categoryRepository.Create(category);
+            var createCategory = await _categoryRepository.Create(category);
 
-            if (result == null)
+            if (createCategory == null)
             {
                 _logger.LogError($"[Dashboard controller] NewCategory() failed, error message: result is null");
                 return NotFound("Category not found");
@@ -210,26 +217,27 @@ public class DashBoardController : Controller
     [Authorize(Roles = "Admin")]
     public async Task<IActionResult> DeleteCategory(int id)
     {
+        // Get the category from the database
         var category = await _categoryRepository.GetTById(id);
 
+        // If the category does not exist, return not found
         if (category == null)
         {
             _logger.LogError($"[Dashboard controller] DeleteCategory() failed, error message: category is null");
             return NotFound("Category not found");
         }
 
-        // Delete the old picture if it exists
-        if (category.PicturePath != null)
+        // If the picture path is not null, delete the file. If the file is in the wwwroot folder, delete it
+        // DeleteFile will return true if the file does not exist
+        if (category.PicturePath != null && !DeleteFile(category.PicturePath))
         {
-            if (!DeleteFile(category.PicturePath))
-            {
-                _logger.LogError($"[Dashboard controller] DeleteCategory() failed, error message: deleteFile failed");
-                return StatusCode(500, "Internal server error, could not delete file");
-            }
+            _logger.LogError($"[Dashboard controller] DeleteCategory() failed, error message: deleteFile failed");
+            return StatusCode(500, "Internal server error, could not delete file");
         }
 
-        var result = await _categoryRepository.Delete(id);
-        if (!result)
+        // Try to delete the category
+        var deleteCategory = await _categoryRepository.Delete(id);
+        if (!deleteCategory)
         {
             _logger.LogError($"[Dashboard controller] DeleteCategory() failed, error message: result is null");
             return NotFound("Category not found");
@@ -238,21 +246,20 @@ public class DashBoardController : Controller
         return RedirectToAction("AdminDashboard");
     }
 
-    public async Task<string> FileUpload(Category category)
+    public async Task<string> FileUpload(IFormFile file)
     {
-        // Get the file
-        var file = Request.Form.Files.FirstOrDefault();
+        // Set the max size of the files
+        long maxSize = 8 * 1024 * 1024; // 8MB
 
         // If the file is null or empty
-        if (file == null || file.Length == 0)
+        if (file.Length == 0)
         {
-            _logger.LogError($"[Dashboard controller] FileUpload() failed, error message: file is null or empty");
+            _logger.LogError($"[Dashboard controller] FileUpload() failed, error message: file is empty");
             return "";
         }
 
-        long maxSize = 8 * 1024 * 1024; // 8MB
-
-        if (file.Length > maxSize) // If the file is greater than 8MB
+        // If the file is greater than 8MB
+        if (file.Length > maxSize)
         {
             _logger.LogError($"[Dashboard controller] FileUpload() failed, error message: file to large");
             return "";
@@ -264,10 +271,16 @@ public class DashBoardController : Controller
         // Create the path to the file
         string filePath = Path.Combine(Directory.GetCurrentDirectory(), "wwwroot", "images", "categories", fileName);
 
+        // Get the directory name
+        string? dirName = Path.GetDirectoryName(filePath);
+
         // If the directory does not exist, create it
-        if (!Directory.Exists(Path.GetDirectoryName(filePath)))
+        if (!Directory.Exists(dirName))
         {
-            Directory.CreateDirectory(Path.GetDirectoryName(filePath));
+            // Try to create the directory if it does not exist
+            if (dirName != null) Directory.CreateDirectory(dirName);
+            _logger.LogError($"[Dashboard controller] FileUpload() failed, error message: could not create directory");
+            return "";
         }
 
         // Copy the file to the path
@@ -275,32 +288,30 @@ public class DashBoardController : Controller
         {
             await file.CopyToAsync(fs);
         }
+        // Source: https://learn.microsoft.com/en-us/dotnet/api/system.io.file.create?view=net-6.0
 
-        return "../images/categories/" + fileName;
+        return Path.Combine("../", "images", "categories", fileName);
     }
 
-    public bool DeleteFile(string path)
+    public bool DeleteFile(string deletePath)
     {
-        string deletePath = path;
-
         //Replace the ../ with wwwroot/ to get the correct path
         deletePath = deletePath.Replace("../", "wwwroot/");
-
-        Console.WriteLine(deletePath);
 
         // Check if the file exists, if i does not exist it's probably a external file 
         if (System.IO.File.Exists(deletePath))
         {
             // Delete the file
             System.IO.File.Delete(deletePath);
-            //https://learn.microsoft.com/en-us/dotnet/api/system.io.file.delete?view=net-7.0#system-io-file-delete(system-string)
 
+            // Check if the file got deleted
             if (System.IO.File.Exists(deletePath))
             {
                 _logger.LogError(
                     $"[Dashboard controller] DeleteCategory() failed, error message: could not delete file");
                 return false;
             }
+            //Source: https://learn.microsoft.com/en-us/dotnet/api/system.io.file.delete?view=net-6.0
         }
 
         return true;
@@ -316,33 +327,33 @@ public class DashBoardController : Controller
         if (ModelState.IsValid)
         {
             // Tries to update tag in repo, logs error if it cannot update tag
-            bool updated = await _tagsRepository.Update(tag);
-            if (!updated)
+            bool updateTag = await _tagsRepository.Update(tag);
+            if (!updateTag)
             {
                 TempData["ErrorMessage"] = "Tag update failed.";
-                _logger.LogWarning("[Dashboard controller] Tag update failed for {@tag}", tag); 
+                _logger.LogWarning("[Dashboard controller] Tag update failed for {@tag}", tag);
             }
         }
 
         TempData["TestMessage"] = "This is a test, babe.";
         // Redirects to admin dashboard
-        return RedirectToAction("AdminDashboard"); 
+        return RedirectToAction("AdminDashboard");
     }
 
     // Post request to create a new tag in the repo and redirect to admin dashboard
     [HttpPost]
     [Authorize(Roles = "Admin")]
-    public async Task<IActionResult> NewTag(Category tag)
+    public async Task<IActionResult> NewTag(Tag tag)
     {
         // Checks if the submitted form values passes validation
         if (ModelState.IsValid)
         {
             // Tries to create tag in repo, logs error if it cannot create tag 
-            var newCategory = await _categoryRepository.Create(tag);
-            if (newCategory == null)
+            var newTag = await _tagsRepository.Create(tag);
+            if (newTag == null)
                 _logger.LogWarning("[Dashboard controller] Tag creation failed for {@tag}", tag);
         }
-        
+
         // Redirects to admin dashboard
         return RedirectToAction("AdminDashboard");
     }
@@ -353,10 +364,10 @@ public class DashBoardController : Controller
     public async Task<IActionResult> DeleteTag(int id)
     {
         // Tries to delete the tag, logs and returns error if there is no category to delete
-        var result = await _tagsRepository.Delete(id);
-        if (!result)
+        var deleteTag = await _tagsRepository.Delete(id);
+        if (!deleteTag)
         {
-            _logger.LogError($"[Dashboard controller] DeleteCategory() failed, error message: result is null");
+            _logger.LogError($"[Dashboard controller] DeleteTag() failed, error message: result is null");
             return NotFound("Category not found");
         }
